+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Add
 - Support for multi-drive in CANopen's NetStatusListener.
 - EtherCAT communication via CoE (SDOs).
+- Add image attribute to dictionary class.
 
 ### Fixed
 - Fix CANopen load_firmware function.